requirements.txt
absl-py==2.3.1
astunparse==1.6.3
audioread==3.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
decorator==5.2.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
joblib==1.5.2
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.2
namex==0.1.0
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
PyYAML==6.0.3
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
